roles/apache: Add task to restart apache2 service
When you install apache2, it is started automatically after installation
process with default configs. including port 80. Therefore, instead of
notifying task to restart the service with the new configurations, we
should manually trigger the handler task to restart the service if
there was any change in the configurations.

I blogged about a similar concept:
https://oguya.ch/linux/2014/12/17/ansible-prompt-handlers

Signed-off-by: James Oguya <oguyajames@gmail.com>